Split area events.
